Add support for ignoring specific subdirectories
Added a new parameter `ignore_subdirs` to the `monitor` function which
allows users to specify a list of subdirectory names that should be
ignored during file monitoring. This feature is useful when certain
subdirectories contain files that are not relevant to the monitoring
process.  The `FileModificationHandler` class has been updated to
respect this new parameter and ignore file changes inside the specified
subdirectories.