Add directory summarization functions
Added a new package called `Node` which contains functions for
generating summaries of directory contents.  It includes the function
`generate_summary()` that uses an AI chatbot to generate summaries and
the function `get_leaf()` that gets the leaf node directory of a given
path.  Additionally, it updates the `File` package to include the new
`show_tree()` function which generates a string representation of a
directory tree.