Add make_result module
Added a new module called `make_result` which contains a function for
dynamically creating dataclass instances from keyword arguments. This
function generates a dataclass based on the provided keyword arguments
and their types, and then returns an instance of that class initialized
with the given values.

The `make_result` function is designed to be used in situations where
custom data classes need to be created at runtime without explicitly
defining them beforehand. This can be useful for creating result objects
with dynamic structures.