Add bounding box support to show function
Added a new parameter `boxes` to the `show` function in
A_GIS.Image.show. This allows users to pass in a list of bounding boxes
to be drawn on the image. Each box is defined by its upper-left and
lower-right corner coordinates, color, and optional line width.

The function now draws these bounding boxes on the image using
matplotlib patches. The line width is automatically calculated if not
provided.